Extract common method to get env template files and ask overwrite
